For a selected directory, add the ability to skip recursive file enumeration.
When a directory is selected, all files within it, recursively, are added to the list of 
files selected.  This is not really what the user intends (i.e. if the files are displayed 
as a tree and the user selects from the first to the last, the user really only wants to 
interact with the selected files, not all those in any selected directory).  

Introduce the AccumRepoFiles class to keep track of files actually selected by the user 
as well as those under any selected directory and provide a way to get either list 
individually or both file lists together.