feat: improve right-click menu for operations on references.
Instead of having a big list of references which can be deleted or renamed in the main popup menu  we add a sub-menu for each kind of operation if there is more than one reference for an operation.  If there is only a single reference for an operation, it is added directly to the menu as per the previous implementation.

There's a pro/con here:
* pro: best use of menu with easiest operation in the most common scenario (a single reference).
* con: menu dynamically changes depending on the data, which means users can't learn muscle-memory. For that it's best to maintain a constant top-level menu and enable/disable menu items as required. This might also make the code a little cleaner.