feat: give the `Rename Branch` dialog a better size.
Ideally I would like the dialog to start with a suitable size based on the current branch name, but not force it to be the minimum. However, I could not find a way to do this.

Secondly, it would be nice if the dialog automatically expanded up to some maximum size as the name gets longer, again, there seems to be no easy way to do this.