Set head/stash reference after pop/drop stash
Fixes the issue  "Dropping multiple stashs in a row segfaults #637".
The cause of the crash was that the commit list was not properly
updated after pop/drop stash actions.  With this change the commit
list is updated or when the stash gets empty (after last pop/drop) 
then the head is selected as current reference.