Set  head reference after pop stash or refresh
When executing "Stash->Pop Stash" from the menu bar or when popping a
specific stash from the commit list then the repo view is set to the head
revision and the commit list is updated with the content from the head
revision. 

Before the when doing a "Refresh" and being on the stash view the commit list 
stayed on the stashes but the head revision was selected. Now the commit list 
contains the commits from the head revision.