feat(whitespace): show whitespace in editor views
This commit adds a new Settings entry "ShowWhitespaceinEditor"
("editor/view/showWhitespace") to the Settings class, and a new checkbox
in the Editor settings dialog.

This setting defaults to off/false, preserving the current behaviour of
the application.

If this setting is toggled to on, whitespace is rendered in the HunkWidget
and BlameEditor views.